Simplify the project layout
See https://github.com/engula/engula/issues/59 for more details.
