+++ src/api/Cargo.toml
@@ -1,7 +1,8 @@
 [package]
-name = "engula-api"
+name = "api"
 version = "0.1.0"
 edition = "2021"
+publish = false
 
 [dependencies]
 datatypes = { path = "datatypes" }
