+++ src/background/src/error.rs
