--- src/engula-framework/Cargo.toml
@@ -1,7 +1,0 @@
-[package]
-name = "engula-framework"
-version = "0.1.0"
-edition = "2021"
-
-[dependencies]
-microunit = { path = "microunit" }
