--- src/engula-kernel/Cargo.toml
@@ -1,8 +1,0 @@
-[package]
-name = "engula-kernel"
-version = "0.1.0"
-edition = "2021"
-
-[dependencies]
-manifest = { path = "manifest" }
-background = { path = "background" }
