--- src/engula-platform/Cargo.toml
@@ -1,7 +1,0 @@
-[package]
-name = "engula-platform"
-version = "0.1.0"
-edition = "2021"
-
-[dependencies]
-storage = { path = "storage" }
