--- src/engula/Cargo.toml
@@ -1,10 +1,0 @@
-[package]
-name = "engula"
-version = "0.1.0"
-edition = "2021"
-
-[dependencies]
-engula-api = { path = "../engula-api" }
-engula-framework = { path = "../engula-framework" }
-clap = "3.0.0-beta.5"
-tokio = { version = "1", features = ["full"] }
