--- src/engula/src/hello_unit.rs
@@ -1,43 +1,0 @@
-// Copyright 2021 The Engula Authors.
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-// http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-use engula_framework::microunit::{async_trait, Result, Unit, UnitBuilder, UnitDesc, UnitSpec};
-
-pub struct HelloUnit {
-    id: String,
-}
-
-#[async_trait]
-impl Unit for HelloUnit {
-    async fn desc(&self) -> UnitDesc {
-        UnitDesc {
-            id: self.id.clone(),
-        }
-    }
-}
-
-#[derive(Default)]
-pub struct HelloUnitBuilder {}
-
-#[async_trait]
-impl UnitBuilder for HelloUnitBuilder {
-    fn kind(&self) -> &str {
-        "hello"
-    }
-
-    async fn spawn(&self, id: String, _spec: UnitSpec) -> Result<Box<dyn Unit>> {
-        let unit = HelloUnit { id };
-        Ok(Box::new(unit))
-    }
-}
