+++ src/manifest/Cargo.toml
