+++ src/manifest/src/lib.rs
