+++ src/manifest/src/version_edit.rs
